requirements.txt
pip==25.2
certifi==2025.8.3
charset_normalizer==3.4.3
click==8.2.1
colorama==0.4.6
colorzero==2.0
contourpy==1.3.3
cycler==0.12.1
daqhats==1.4.1.0
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
et-xmlfile==2.0.0
fonttools==4.59.0
future==1.0.0
gpiozero==2.0.1
hightime==0.2.2
idna==3.10
kiwisolver==1.4.9
matplotlib==3.10.5
nidaqmx==1.2.0
numexpr==2.11.0
numpy==2.3.2
odfpy==1.4.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pandastable==0.14.0
pigpio==1.78
pillow==11.3.0
pyparsing==3.2.3
pyserial==3.5
python-calamine==0.4.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
pyxlsb==1.0.10
requests==2.32.4
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
xlrd==2.0.2
xlsxwriter==3.2.5